refactor cli wrapper client surfaces
